Fix test warning
An additional information instance was already assigned to a regulation
instance in the factory fixture, which invalidated the additional information
collection of a regulation instance. It was then appended again to the
collection. This change removes the assignment of additional information
to a regulation in the factory fixture and returns the additional information
instance as uncommitted. It is then the user's responsibility to add the
additional information instance to a session.